+++ tests/Constraints/Fails.php
@@ -26,8 +26,6 @@ class Fails extends Constraint
      */
     public function __construct(Factory $factory, array $rules)
     {
-        parent::__construct();
-
         $this->factory = $factory;
         $this->rules = $rules;
     }