+++ tests/Unit/PostalCodeForValidationTest.php
@@ -18,7 +18,7 @@ class PostalCodeForValidationTest extends ValidationTest
     /**
      * {@inheritDoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();