+++ tests/Unit/PostalCodeValidationTest.php
@@ -10,7 +10,7 @@ class PostalCodeValidationTest extends ValidationTest
     /**
      * {@inheritDoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();