+++ tests/Unit/ValidationTest.php
@@ -20,7 +20,7 @@ abstract class ValidationTest extends TestCase
     /**
      * {@inheritDoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();